update to 0.14b
http://elm-chan.org/fsw/ff/arc/ff14b.zip

Linked from: http://elm-chan.org/fsw/ff/00index_e.html

Previously we manually patched ff.c to include sdmmc_diskio.h, but
this is not necessary and not advisable: the diskio.h included in the
project is what FatFs expects, and deviations could lead to bugs.

ffconf.h was modified to match the previous configuration.